Rename module `sys-util` -> `webapp-util`.
