+++ doc/quick-start/code/index.mjs
@@ -6,7 +6,7 @@ import { default as CONFIG } from './config-framework.mjs';
 import { Host, KeepRunning } from '@this/host';
 import { Loggy } from '@this/loggy';
 import { Warehouse } from '@this/sys-framework';
-import { BaseSystem } from '@this/sys-util';
+import { BaseSystem } from '@this/webapp-util';
 
 
 /**
