--- src/sys-util/README.md
@@ -1,12 +1,0 @@
-@this/sys-util
-==============
-
-Higher level utility functionality in support of the Lactoserv framework. These
-classes aren't core to the framework's operation but are useful when building
-stuff on top of it.
-
-- - - - - - - - - -
-```
-Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
-SPDX-License-Identifier: Apache-2.0
-```
