+++ src/webapp-builtins/export/ProcessIdFile.js
@@ -9,7 +9,7 @@ import { WallClock } from '@this/clocks';
 import { Duration } from '@this/data-values';
 import { Statter } from '@this/fs-util';
 import { ProcessUtil } from '@this/host';
-import { BaseFileService } from '@this/sys-util';
+import { BaseFileService } from '@this/webapp-util';
 import { MustBe } from '@this/typey';
