+++ src/webapp-builtins/export/ProcessInfoFile.js
@@ -9,7 +9,7 @@ import { Duration } from '@this/data-values';
 import { Statter } from '@this/fs-util';
 import { Host, ProcessInfo, ProcessUtil, ProductInfo }
   from '@this/host';
-import { BaseFileService, Saver } from '@this/sys-util';
+import { BaseFileService, Saver } from '@this/webapp-util';
 import { MustBe } from '@this/typey';
