+++ src/webapp-builtins/export/SyslogToFile.js
@@ -5,7 +5,7 @@ import { EventTracker, LinkedEvent } from '@this/async';
 import { WallClock } from '@this/clocks';
 import { Duration } from '@this/data-values';
 import { Loggy, TextFileSink } from '@this/loggy';
-import { BaseFileService, Rotator } from '@this/sys-util';
+import { BaseFileService, Rotator } from '@this/webapp-util';
 import { MustBe } from '@this/typey';
