+++ src/webapp-builtins/package.json
@@ -24,7 +24,7 @@
     "@this/net-protocol": "*",
     "@this/net-util": "*",
     "@this/sys-framework": "*",
-    "@this/sys-util": "*",
+    "@this/webapp-util": "*",
     "@this/typey": "*"
   }
 }
