+++ src/webapp-util/README.md
@@ -1,0 +1,12 @@
+@this/webapp-util
+=================
+
+Higher level utility functionality in support of the web application server
+framework. These classes aren't core to the framework's operation but are useful
+when building stuff on top of it.
+
+- - - - - - - - - -
+```
+Copyright 2022-2024 the Lactoserv Authors (Dan Bornstein et alia).
+SPDX-License-Identifier: Apache-2.0
+```
