+++ src/webapp-util/export/Rotator.js
